change layout and base css of project
