+++ src/index.css
@@ -3,9 +3,9 @@
   line-height: 1.5;
   font-weight: 400;
 
-  color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
+  color-scheme: light;
+  color: #535931;
+  background-color: #d7dcbc;
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
@@ -15,11 +15,11 @@
 
 a {
   font-weight: 500;
-  color: #646cff;
+  color: #a88467;
   text-decoration: inherit;
 }
 a:hover {
-  color: #535bf2;
+  color: #a88467; /*maybe change that later*/
 }
 
 body {
@@ -36,33 +36,23 @@ h1 {
 }
 
 button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
+  border-radius: 20px;
+  border: 3px solid transparent;
+  padding: 14px 28px;
+  font-size: 22px;
+  font-weight: 700;
   font-family: inherit;
-  background-color: #1a1a1a;
+  background-color: #b5bf84;
+  color: white;
   cursor: pointer;
-  transition: border-color 0.25s;
+  transition: all 0.5s ease;
 }
 button:hover {
-  border-color: #646cff;
+  border-color: #535931;
+  border-width: 3px;
+  color: #535931;
 }
 button:focus,
 button:focus-visible {
   outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-  :root {
-    color: #213547;
-    background-color: #ffffff;
-  }
-  a:hover {
-    color: #747bff;
-  }
-  button {
-    background-color: #f9f9f9;
-  }
 }
